back: handle missing `Permutated.csv` case
Needed to be handled in two locations

1. in `file_io.write()`
2. in `__main__.verify()`

+++ back/src/chatbot_util/file_io.py
@@ -201,7 +201,9 @@ def write(
     nums: dict[str, int],
 ) -> None:
     """Format questions and topics, write to csv file"""
-    os.rename(FILENAMES["permutated"], FILENAMES["permutated"] + ".backup")
+    if os.path.exists(FILENAMES["permutated"]):
+        os.rename(FILENAMES["permutated"], FILENAMES["permutated"] + ".backup")
+
     with open(FILENAMES["permutated"], "w", encoding="utf-8") as csvfile:
         csvfile.write('"question","answer"\n')
         indices = {
